requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
google-generativeai==0.8.3
neo4j==5.14.1
graphdatascience==1.9.0
python-dotenv==1.0.0
langchain==0.1.0
langgraph==0.0.26
langchain-google-genai==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
